Fixed some incorrect data types

+++ src/drivers/uart/Uart_public.h
@@ -147,7 +147,7 @@ ErrorCode Uart_init_specific(Uart_DeviceId uart_id_in);
 ErrorCode Uart_send_bytes(
     Uart_DeviceId uart_id_in,
     uint8_t *p_data_in, 
-    size_t length_in
+    uint32_t length_in
 );
 
 /**
@@ -164,7 +164,7 @@ ErrorCode Uart_send_bytes(
 ErrorCode Uart_recv_bytes(
     Uart_DeviceId uart_id_in,
     uint8_t *p_data_out,
-    size_t length_in
+    uint32_t length_in
 );
 
 /**
